[rename] signUpSchema.ts -> schema.ts
