+++ apps/fe/src/app/register/schema.ts
